Fix attribute copying (#117)
* Make CodapAttribute type looser

* Copy properties of attributes regardless of type

* Remove console.log
